Uint256

+++ src/test/mgr-unit.t.sol
@@ -243,7 +243,7 @@ contract TinlakeManagerUnitTest is DSTest, DSMath {
     function wipe(uint wad) public {
         uint selfBalanceDAI = dai.balanceOf(self);
         uint totalSupplyDAI = dai.totalSupply();
-       
+
         mgr.wipe(wad);
 
         // check DAI were transferred & burned
@@ -269,8 +269,8 @@ contract TinlakeManagerUnitTest is DSTest, DSMath {
 
 
         // check allowance set for dai & collateral
-        assertEq(dai.allowance(mgr_, newMgr_), uint(-1));
-        assertEq(drop.allowance(mgr_, newMgr_), uint(-1));
+        assertEq(dai.allowance(mgr_, newMgr_), uint256(-1));
+        assertEq(drop.allowance(mgr_, newMgr_), uint256(-1));
         // assert live is set to false
         assert(!mgr.live());
     }
@@ -283,7 +283,7 @@ contract TinlakeManagerUnitTest is DSTest, DSMath {
         (, uint256 rate, , ,) = vat.ilks(ilk);
         // assert correct DAI amount was written off
         uint tab = mul(rate, art);
-        
+
         assertEq(mgr.tab(), tab);
         // assert sink called
         assert(!mgr.glad());
@@ -360,7 +360,7 @@ contract TinlakeManagerUnitTest is DSTest, DSMath {
     }
 
     function testFailDrawAboveCeiling(uint wad) public {
-        assert(ceiling < wad); 
+        assert(ceiling < wad);
         testLock();
         draw(wad);
     }
@@ -436,7 +436,7 @@ contract TinlakeManagerUnitTest is DSTest, DSMath {
         tell();
         sink();
     }
-    
+
     function testFailSinkGlobalSettlement() public {
         uint wad = 100 ether;
         testDraw(wad);
